Dataset with real time and day ahead prices
I created d-1 and d-7 columns for real-time, day ahead, and actual load columns. I realized that not every day has all the hours (like 2022-09-11 is missing the hour 05:00 etc.), so subtracting 24 hours or 24*7 hours from the index for d-1 and d-7 did not work. I worked on it a little bit.